requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
iniconfig==1.1.1
packaging==21.3
paramiko==2.9.2
Pint==0.18
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyfrc==2023.0.1
PyNaCl==1.5.0
pynetconsole==2.0.2
pynetworktables==2021.0.0
pyntcore==2023.3.2.1
pyparsing==3.0.7
pytest==6.2.5
pytest-reraise==2.1.2
robotpy==2023.3.2.1
robotpy-apriltag==2023.3.2.1
robotpy-commands-v2==2023.3.2.1
robotpy-cscore==2023.3.2.1
robotpy-ctre==2023.0.0
robotpy-hal==2023.3.2.1
robotpy-halsim-ds-socket==2023.3.2.1
robotpy-halsim-gui==2023.3.2.1
robotpy-halsim-ws==2023.3.2.1
robotpy-installer==2023.0.1
robotpy-navx==2023.0.3
robotpy-photonvision==2023.2.2
robotpy-rev==2023.1.3
robotpy-wpilib-utilities==2023.0.0
robotpy-wpimath==2023.3.2.1
robotpy-wpinet==2023.4.3.0
robotpy-wpiutil==2023.3.2.1
six==1.16.0
toml==0.10.2
wpilib==2023.3.2.1